Refactore folder structure #108
