+++ norbit-mobileapp/flutter_application_1/lib/blocs/device_name_bloc.dart
