+++ norbit-mobileapp/flutter_application_1/lib/blocs/username_bloc.dart
