+++ norbit-mobileapp/flutter_application_1/lib/main.dart
@@ -2,18 +2,17 @@ import 'package:flutter/material.dart';
 import 'package:flutter_application_1/screens/signin_screen.dart';
 import 'package:flutter_application_1/services/login_service.dart';
 import 'package:flutter_application_1/services/mqtt_service.dart';
-import 'package:flutter_application_1/services/location_service.dart';
-import 'package:flutter_application_1/services/noise_service.dart';
-import 'package:flutter_application_1/services/lux_service.dart';
-import 'package:flutter_application_1/services/accelerometer_service.dart';
-import 'blocs/connectivity/device_name_bloc.dart';
-import 'blocs/connectivity/location_bloc.dart';
-import 'blocs/connectivity/token_bloc.dart';
-import 'blocs/connectivity/username_bloc.dart';
-import 'screens/home_screen.dart';
-import '../blocs/connectivity/noise_bloc.dart';
-import '../blocs/connectivity/lux_bloc.dart';
-import '../blocs/connectivity/accelerometer_bloc.dart';
+import 'package:flutter_application_1/services/sensors/location_service.dart';
+import 'package:flutter_application_1/services/sensors/noise_service.dart';
+import 'package:flutter_application_1/services/sensors/lux_service.dart';
+import 'package:flutter_application_1/services/sensors/accelerometer_service.dart';
+import 'blocs/device_name_bloc.dart';
+import 'blocs/sensors/location_bloc.dart';
+import 'blocs/token_bloc.dart';
+import 'blocs/username_bloc.dart';
+import '../blocs/sensors/noise_bloc.dart';
+import '../blocs/sensors/lux_bloc.dart';
+import '../blocs/sensors/accelerometer_bloc.dart';
 import 'blocs/start_stop_bloc.dart';
 import 'package:provider/provider.dart';
