+++ norbit-mobileapp/flutter_application_1/lib/screens/home_screen.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/location_bloc.dart';
-import 'package:flutter_application_1/services/noise_service.dart';
-import 'package:flutter_application_1/services/lux_service.dart';
-import 'package:flutter_application_1/services/accelerometer_service.dart';
+import 'package:flutter_application_1/blocs/sensors/location_bloc.dart';
+import 'package:flutter_application_1/services/sensors/noise_service.dart';
+import 'package:flutter_application_1/services/sensors/lux_service.dart';
+import 'package:flutter_application_1/services/sensors/accelerometer_service.dart';
 import 'package:provider/provider.dart';
-import '../blocs/connectivity/noise_bloc.dart';
-import '../blocs/connectivity/lux_bloc.dart';
-import '../blocs/connectivity/accelerometer_bloc.dart';
+import '../blocs/sensors/noise_bloc.dart';
+import '../blocs/sensors/lux_bloc.dart';
+import '../blocs/sensors/accelerometer_bloc.dart';
 import '../blocs/start_stop_bloc.dart';
-import '../services/location_service.dart';
+import '../services/sensors/location_service.dart';
 import '../widgets/sensors/accelerometer_widget.dart';
 import '../widgets/sensors/location_widget.dart';
 import '../widgets/sensors/lux_widget.dart';
