+++ norbit-mobileapp/flutter_application_1/lib/widgets/device_data.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/device_name_bloc.dart';
+import 'package:flutter_application_1/blocs/device_name_bloc.dart';
 
 class DeviceData extends StatelessWidget {
   final DeviceNameBloc deviceNameBloc;
