+++ norbit-mobileapp/flutter_application_1/lib/widgets/device_popup.dart
@@ -2,10 +2,9 @@ import 'dart:convert';
 import 'package:amplify_core/amplify_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:provider/provider.dart';
-import '../blocs/connectivity/device_name_bloc.dart';
-import '../blocs/connectivity/token_bloc.dart';
-import '../blocs/connectivity/username_bloc.dart';
+import '../blocs/device_name_bloc.dart';
+import '../blocs/token_bloc.dart';
+import '../blocs/username_bloc.dart';
 import '../services/aws_service.dart';
 import '../services/save_service.dart';
