+++ norbit-mobileapp/flutter_application_1/lib/widgets/sensors/accelerometer_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/accelerometer_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/accelerometer_bloc.dart';
 import 'package:sensors_plus/sensors_plus.dart';
 
 class AccelerometerWidget extends StatelessWidget {
