+++ norbit-mobileapp/flutter_application_1/lib/widgets/sensors/lux_widget.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/lux_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/lux_bloc.dart';
 
 class LuxWidget extends StatelessWidget {
   final LuxBloc luxBloc;
