+++ norbit-mobileapp/flutter_application_1/lib/widgets/sidebar_widget.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/device_name_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/username_bloc.dart';
+import 'package:flutter_application_1/blocs/device_name_bloc.dart';
+import 'package:flutter_application_1/blocs/username_bloc.dart';
 import 'package:flutter_application_1/screens/signin_screen.dart';
 import 'package:provider/provider.dart';
-import '../blocs/connectivity/token_bloc.dart';
+import '../blocs/token_bloc.dart';
 import '../services/login_service.dart';
 import 'device_data.dart';
 import 'device_popup.dart';
