+++ norbit-mobileapp/flutter_application_1/test/lux_widget_test.dart
@@ -1,8 +1,7 @@
-// Import the required packages
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_application_1/widgets/sensors/lux_widget.dart';
-import 'package:flutter_application_1/blocs/connectivity/lux_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/lux_bloc.dart';
 import 'package:provider/provider.dart';
 
 void main() {
