+++ norbit-mobileapp/flutter_application_1/test/noise_bloc_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:flutter_application_1/blocs/connectivity/noise_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/noise_bloc.dart';
 
 void main() {
   group('NoiseBloc', () {
