+++ norbit-mobileapp/flutter_application_1/test/start_stop_button_widget_test.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/blocs/connectivity/accelerometer_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/device_name_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/location_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/lux_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/noise_bloc.dart';
-import 'package:flutter_application_1/blocs/connectivity/username_bloc.dart';
-import 'package:flutter_application_1/services/accelerometer_service.dart';
-import 'package:flutter_application_1/services/location_service.dart';
-import 'package:flutter_application_1/services/lux_service.dart';
-import 'package:flutter_application_1/services/noise_service.dart';
+import 'package:flutter_application_1/blocs/sensors/accelerometer_bloc.dart';
+import 'package:flutter_application_1/blocs/device_name_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/location_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/lux_bloc.dart';
+import 'package:flutter_application_1/blocs/sensors/noise_bloc.dart';
+import 'package:flutter_application_1/blocs/username_bloc.dart';
+import 'package:flutter_application_1/services/sensors/accelerometer_service.dart';
+import 'package:flutter_application_1/services/sensors/location_service.dart';
+import 'package:flutter_application_1/services/sensors/lux_service.dart';
+import 'package:flutter_application_1/services/sensors/noise_service.dart';
 import 'package:flutter_application_1/services/mqtt_service.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:provider/provider.dart';
